Optimizer: Fix (?:abc)+abc optimization in combine-repeating-patterns
